Move selection models to lib/models/
