+++ lib/models/composite-list-selection.js
