+++ lib/models/file-patch-selection.js
