+++ lib/models/list-selection.js
