+++ lib/views/file-patch-view.js
@@ -8,7 +8,7 @@ import {autobind} from 'core-decorators';
 import HunkView from './hunk-view';
 import SimpleTooltip from './simple-tooltip';
 import Commands, {Command} from './commands';
-import FilePatchSelection from './file-patch-selection';
+import FilePatchSelection from '../models/file-patch-selection';
 import Switchboard from '../switchboard';
 import RefHolder from '../models/ref-holder';
