+++ lib/views/staging-view.js
@@ -13,7 +13,7 @@ import isEqual from 'lodash.isequal';
 
 import FilePatchListItemView from './file-patch-list-item-view';
 import MergeConflictListItemView from './merge-conflict-list-item-view';
-import CompositeListSelection from './composite-list-selection';
+import CompositeListSelection from '../models/composite-list-selection';
 import ResolutionProgress from '../models/conflicts/resolution-progress';
 import ModelObserver from '../models/model-observer';
 import FilePatchController from '../controllers/file-patch-controller';
